+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - PodPolicy `resourceRequirements` field is renamed to `resources`
 - Default etcd version is changed to `3.1.2`
+- Self-hosted etcd pod uses hostPath with path `/var/etcd/$ns-$member`
 
 ### Removed